test(resolve): move bun-store tests to their own file (2000-line lint cap)
resolve/tests.rs was already ~1939 lines; appending the two #5914
regression tests pushed it past the workspace's 2000-line file-size
lint gate. Move them to a new resolve/bun_store_tests.rs sibling
module, matching the existing pattern of splitting resolve.rs's
concerns into focused files (native_library.rs, tsconfig_paths.rs).
No behavior change.
