+++ crates/perry/src/commands/compile/resolve.rs
@@ -122,6 +122,8 @@ pub fn find_perry_workspace_root() -> Option<PathBuf> {
 }
 
 #[cfg(test)]
+mod bun_store_tests;
+#[cfg(test)]
 mod tests;
 
 /// Packages that Perry provides built-in native extensions for.
